pyproject.toml
[project]
name = "ie_visa_bot"
version = "0.1.0"
description = "Python project using uv"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "beautifulsoup4",
    "pandas",
    "pyexcel-ods",
    "openpyxl",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
pip = true
use-lockfile = true